// turismo-frontend/src/app/core/models/user.model.ts
export interface User {
  id?: number;
  name: string;
  email: string;
  phone: string;
  active?: boolean;
  created_at?: string;
  updated_at?: string;
  google_id?: string;
  foto_perfil?: string;
  foto_perfil_url?: string | null;
  avatar?: string | null;
  email_verified_at?: string | null;
  roles?: Array<{
    id: number;
    name: string;
    permissions?: string[] | Permission[];
  }>;
  country?: string;
  birth_date?: string;
  address?: string;
  gender?: string;
  preferred_language?: string;
  last_login?: string;
  permissions?: Permission[];
  pivot?: {
    user_id?: number;
    emprendedor_id?: number;
    es_principal?: boolean;
    rol?: string;
    created_at?: string;
    updated_at?: string;
  };
}

export interface Role { 
  id: number;
  name: string;
  guard_name?: string;
  created_at?: string;
  updated_at?: string;
  permissions?: string[] | Permission[];
  pivot?: any;
}

export interface Permission {
  id: number;
  name: string;
  guard_name: string;
  created_at?: string;
  updated_at?: string;
  pivot?: any;
}